% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ransac.R
\name{higher.consensus}
\alias{higher.consensus}
\title{Choose model with higher consensus set}
\usage{
higher.consensus(family.fun, el.error, el.model, best.error, best.model)
}
\arguments{
\item{best.error}{}
}
\value{

}
\description{
Choose model with higher consensus set
}
