% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ransac.R
\name{lower.loss}
\alias{lower.loss}
\title{Choose model with lower loss}
\usage{
lower.loss(el.error, el.model, best.error, best.model)
}
\arguments{
\item{best.error}{}
}
\value{

}
\description{
Choose model with lower loss
}
