% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.glmnet.R
\name{ransac.binomial.glmnet}
\alias{ransac.binomial.glmnet}
\title{Support glmnet logistic regression}
\usage{
ransac.binomial.glmnet(auc = F, residuals = "deviance")
}
\description{
Uses the binomial distribution
}

