% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ransac_family.R
\name{ransac.family}
\alias{ransac.family}
\title{Get RANSAC family function}
\usage{
ransac.family(family)
}
\arguments{
\item{family}{}
}
\value{
a ransac.family function
}
\description{
Get RANSAC family function
}
\examples{
ransac.family('ransac.binomial.glm')
}

