% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ransac.R
\name{ransac}
\alias{ransac}
\title{RANSAC}
\usage{
ransac(xdata, ydata, n, threshold, good.fit.perct, k = 100,
  family = "binomial", mc.cores = 1, ...)
}
\arguments{
\item{xdata}{co-variates (rows: observations | coluns: co-variates)}

\item{ydata}{response}

\item{n}{minimum number of observations that will be fitted in the model}

\item{threshold}{error threshold to determine if observation is inliner}

\item{good.fit.perct}{discard models that have less than this percentage of inliners}

\item{k}{number of times model is run}

\item{family}{type of model that should be used}

\item{mc.cores}{number of processes to process the iteration of ransac}
}
\description{
RANSAC
}

