% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_ransac_results.R
\name{plot.ransac}
\alias{plot.ransac}
\title{Function to print the results from RANSAC}
\usage{
\method{plot}{ransac}(result.ransac, xdata, ydata, ydata.prob = NULL,
  family = "binomial.glm", name = "", baseline = list(), show.title = T,
  only_consensus = T, outliers = NULL, show.misclass = T,
  print.plots = T, ...)
}
\arguments{
\item{...}{}
}
\description{
Function to print the results from RANSAC
}
