% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_ransac_results.R
\name{plot.ransac}
\alias{plot.ransac}
\title{Function to print the results from RANSAC}
\usage{
\method{plot}{ransac}(results.ransac, xdata, ydata, family = "binomial",
  name = "", ydata.original = NULL, show.title = T, only_consensus = T,
  ...)
}
\arguments{
\item{...}{}
}
\description{
Function to print the results from RANSAC
}

